+++ tests/xharness/MakefileGenerator.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
-using Xharness;
+using Xharness.Targets;
 
 namespace Xharness
 {
