+++ tests/xharness/Targets/iOSTarget.cs
@@ -2,7 +2,7 @@ using System.IO;
 using System.Xml;
 using Xharness.Utilities;
 
-namespace Xharness
+namespace Xharness.Targets
 {
 	// iOS here means Xamarin.iOS, not iOS as opposed to tvOS/watchOS.
 	public class iOSTarget : Target
