+++ tests/xharness/xharness.csproj
@@ -86,6 +86,7 @@
     <Compile Include="Jenkins\TestTasks\RunTestTask.cs" />
     <Compile Include="Jenkins\TestTasks\RunXITask.cs" />
     <Compile Include="Jenkins\TestTasks\RunXtroTask.cs" />
+    <Compile Include="Targets\iOSTarget.cs" />
     <Compile Include="TestExecutingResult.cs" />
     <Compile Include="TestPlatform.cs" />
     <Compile Include="Jenkins\TestTasks\TestTask.cs" />
@@ -95,12 +96,11 @@
     <Compile Include="Harness.cs" />
     <Compile Include="ProjectFileExtensions.cs" />
     <Compile Include="Utilities\PlistExtensions.cs" />
-    <Compile Include="Utilities\StringExtensions.cs" />
     <Compile Include="Utilities\StringUtils.cs" />
-    <Compile Include="WatchOSTarget.cs" />
-    <Compile Include="TVOSTarget.cs" />
-    <Compile Include="Target.cs" />
-    <Compile Include="UnifiedTarget.cs" />
+    <Compile Include="Targets\WatchOSTarget.cs" />
+    <Compile Include="Targets\TVOSTarget.cs" />
+    <Compile Include="Targets\Target.cs" />
+    <Compile Include="Targets\UnifiedTarget.cs" />
     <Compile Include="AppRunner.cs" />
     <Compile Include="..\..\tools\mtouch\SdkVersions.cs">
       <Link>SdkVersions.cs</Link>
@@ -108,15 +108,14 @@
     <Compile Include="DeviceLogCapturer.cs" />
     <Compile Include="MakefileGenerator.cs" />
     <Compile Include="SolutionGenerator.cs" />
-    <Compile Include="MacTarget.cs" />
+    <Compile Include="Targets\MacTarget.cs" />
     <Compile Include="Jenkins\Jenkins.cs" />
     <Compile Include="Process_Extensions.cs" />
     <Compile Include="Simulators.cs" />
     <Compile Include="TestProject.cs" />
     <Compile Include="GitHub.cs" />
     <Compile Include="Extensions.cs" />
-    <Compile Include="TodayExtensionTarget.cs" />
-    <Compile Include="iOSTarget.cs" />
+    <Compile Include="Targets\TodayExtensionTarget.cs" />
     <Compile Include="MonoNativeInfo.cs" />
     <Compile Include="XmlResultParser.cs" />
     <Compile Include="BCLTestImporter\BCLTestAssemblyDefinition.cs" />
